LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-content-alignment-overflow-002-expected.txt

Grid container width of 60px not enough for 2 column tracks of 50px.
Content-Alignment: center and Overflow-Alignment: default

Grid container height of 150px not enough for 2 row tracks of 100px.
Content-Alignment: center and Overflow-Alignment: unsafe

Grid container width of 60px not enough for 2 column tracks of 50px.
Content-Alignment: center and Overflow-Alignment: safe

Grid container height of 150px not enough for 2 row tracks of 100px.
Content-Alignment: center and Overflow-Alignment: safe

Grid container width of 60px not enough for 2 column tracks of 50px.
Content-Alignment: end and Overflow-Alignment: default

Grid container height of 150px not enough for 2 row tracks of 100px.
Content-Alignment: end and Overflow-Alignment: unsafe

Grid container width of 60px not enough for 2 column tracks of 50px.
Content-Alignment: end and Overflow-Alignment: safe

Grid container height of 150px not enough for 2 row tracks of 100px.
Content-Alignment: end and Overflow-Alignment: safe

Grid container indefinite size using fit-content, hence no possible overflow.
Content-Alignment: end and Overflow-Alignment: unsafe

Grid container indefinite size using fit-content, hence no possible overflow.
Content-Alignment: center and Overflow-Alignment: unsafe

Grid container indefinite size using fit-content, but max-size constraints implies that the content-sized tracks overflow.
Content-Alignment: end and Overflow-Alignment: unsafe

Grid container indefinite size using fit-content, but max-size constraints implies that the content-sized tracks overflow.
Content-Alignment: center and Overflow-Alignment: unsafe

Grid container indefinite size using fit-content, but min-size constraints implies that the content-sized tracks don't overflow.
Content-Alignment: end and Overflow-Alignment: unsafe

Grid container indefinite size using fit-content, but min-size constraints implies that the content-sized tracks don't overflow.
Content-Alignment: center and Overflow-Alignment: unsafe


FAIL .grid 1 assert_equals:
<div class="container" style="margin-bottom: 50px; margin-right: 25px;">
  <div class="grid overflowWidth contentCenter" data-expected-width="60" data-expected-height="300" data-expected-scroll-width="110" data-expected-scroll-height="300">
    <div class="firstRowFirstColumn" data-offset-x="-20" data-offset-y="50" data-expected-width="50" data-expected-height="100"></div>
    <div class="secondRowFirstColumn" data-offset-x="-20" data-offset-y="150" data-expected-width="50" data-expected-height="100"></div>
    <div class="firstRowSecondColumn" data-offset-x="30" data-offset-y="50" data-expected-width="50" data-expected-height="100"></div>
    <div class="secondRowSecondColumn" data-offset-x="30" data-offset-y="150" data-expected-width="50" data-expected-height="100"></div>
  </div>
</div>
scrollWidth expected 110 but got 130
FAIL .grid 2 assert_equals:
<div class="container" style="margin-bottom: 50px; margin-right: 25px;">
  <div class="grid overflowHeight contentCenterUnsafe" data-expected-width="200" data-expected-height="150" data-expected-scroll-width="200" data-expected-scroll-height="205">
    <div class="firstRowFirstColumn" data-offset-x="50" data-offset-y="-25" data-expected-width="50" data-expected-height="100"></div>
    <div class="secondRowFirstColumn" data-offset-x="50" data-offset-y="75" data-expected-width="50" data-expected-height="100"></div>
    <div class="firstRowSecondColumn" data-offset-x="100" data-offset-y="-25" data-expected-width="50" data-expected-height="100"></div>
    <div class="secondRowSecondColumn" data-offset-x="100" data-offset-y="75" data-expected-width="50" data-expected-height="100"></div>
  </div>
</div>
scrollHeight expected 205 but got 230
PASS .grid 3
PASS .grid 4
FAIL .grid 5 assert_equals:
<div class="container" style="margin-bottom: 50px; margin-right: 25px;">
  <div class="grid overflowWidth contentEnd" data-expected-width="60" data-expected-height="300" data-expected-scroll-width="60" data-expected-scroll-height="300">
    <div class="firstRowFirstColumn" data-offset-x="-70" data-offset-y="70" data-expected-width="50" data-expected-height="100"></div>
    <div class="secondRowFirstColumn" data-offset-x="-70" data-offset-y="170" data-expected-width="50" data-expected-height="100"></div>
    <div class="firstRowSecondColumn" data-offset-x="-20" data-offset-y="70" data-expected-width="50" data-expected-height="100"></div>
    <div class="secondRowSecondColumn" data-offset-x="-20" data-offset-y="170" data-expected-width="50" data-expected-height="100"></div>
  </div>
</div>
scrollWidth expected 60 but got 160
FAIL .grid 6 assert_equals:
<div class="container" style="margin-bottom: 50px; margin-right: 25px;">
  <div class="grid overflowHeight contentEndUnsafe" data-expected-width="200" data-expected-height="150" data-expected-scroll-width="200" data-expected-scroll-height="150">
    <div class="firstRowFirstColumn" data-offset-x="70" data-offset-y="-80" data-expected-width="50" data-expected-height="100"></div>
    <div class="secondRowFirstColumn" data-offset-x="70" data-offset-y="20" data-expected-width="50" data-expected-height="100"></div>
    <div class="firstRowSecondColumn" data-offset-x="120" data-offset-y="-80" data-expected-width="50" data-expected-height="100"></div>
    <div class="secondRowSecondColumn" data-offset-x="120" data-offset-y="20" data-expected-width="50" data-expected-height="100"></div>
  </div>
</div>
scrollHeight expected 150 but got 230
PASS .grid 7
PASS .grid 8
PASS .grid 9
PASS .grid 10
FAIL .grid 11 assert_equals:
<div class="container" style="margin-bottom: 100px; margin-right: 25px;">
  <div class="grid contentSizedTracks fit-content maxSize contentEndUnsafe" data-expected-width="100" data-expected-height="100" data-expected-scroll-width="100" data-expected-scroll-height="100">
    <div class="item1 firstRowSecondColumn" data-offset-x="20" data-offset-y="-180" data-expected-width="50" data-expected-height="150"></div>
    <div class="item2 secondRowFirstColumn" data-offset-x="-130" data-offset-y="-30" data-expected-width="150" data-expected-height="100"></div>
  </div>
</div>
scrollWidth expected 100 but got 260
FAIL .grid 12 assert_equals:
<div class="container" style="margin-bottom: 100px; margin-right: 75px;">
  <div class="grid contentSizedTracks fit-content maxSize contentCenterUnsafe" data-expected-width="100" data-expected-height="100" data-expected-scroll-width="180" data-expected-scroll-height="205">
    <div class="item1 firstRowSecondColumn" data-offset-x="100" data-offset-y="-75" data-expected-width="50" data-expected-height="150"></div>
    <div class="item2 secondRowFirstColumn" data-offset-x="-50" data-offset-y="75" data-expected-width="150" data-expected-height="100"></div>
  </div>
</div>
scrollWidth expected 180 but got 230
PASS .grid 13
PASS .grid 14
